+++ helper/User.js
@@ -6,6 +6,9 @@ export const getAllUsers = async () => {
 };
 
 export const getUserByID = async (userId) => {
+  if (!userId) {
+    return console.error("User ID not found");
+  }
   const response = await api.get(`/user/${userId}`);
   return response;
 };
@@ -21,11 +24,17 @@ export const getUserByRole = async () => {
 };
 
 export const updateUser = async (userId, formData) => {
+  if ((!userId, !formData)) {
+    return console.error("Error getting UserID and formdata");
+  }
   const response = await api.put(`/updateuser/${userId}`, formData);
   return response;
 };
 
 export const deleteUser = async (id) => {
+  if (!id) {
+    return console.error("User ID not found");
+  }
   const response = await api.delete(`/deleteuser/${id}`);
   return response;
 };
